menu refactor and social links enums added
