+++ components/footer.component.tsx
@@ -10,17 +10,17 @@ export default function Footer() {
         {/* footer contact */}
         <div className='xl:w-2/4 flex flex-col items-center'>
           <p className='text-lm_icons text-lg text-center font-semibold mb-6 mx-8'>
-          Business & Partnership Proposals
+            Business & Partnership Proposals
           </p>
           <div className='flex justify-around xl:w-[57%] pt-4'>
-            <SocialIcon iconForm={IconForm.Big} iconRef='m' link='https://medium.com/@THEHUB_DAO' alt='medium link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='i' link='https://www.instagram.com/thehub_dao/' alt='instagram link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='l' link='https://www.linkedin.com/company/the-hub-dao/' alt='linkedin link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='w' link='https://twitter.com/THE_HUB_DAO' alt='twitter link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='t' link='https://t.me/thehub_dao' alt='telegram link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='a' link='https://discord.gg/J35NGdPWgq' alt='discord link' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='M' link='mailto:info@thedac.info' alt='The Hub DAO email' />
-            <SocialIcon iconForm={IconForm.Big} iconRef='e' link='https://etherscan.io/token/0x8765b1a0eb57ca49be7eacd35b24a574d0203656' alt='etherscan link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='m' link={ExternalLink.Medium} alt='medium link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='i' link={ExternalLink.Instagram} alt='instagram link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='l' link={ExternalLink.Linkedin} alt='linkedin link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='w' link={ExternalLink.Twitter} alt='twitter link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='t' link={ExternalLink.Telegram} alt='telegram link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='a' link={ExternalLink.Discord} alt='discord link' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='M' link={ExternalLink.MailTo} alt='The Hub DAO email' />
+            <SocialIcon iconForm={IconForm.Regular} iconRef='e' link={ExternalLink.ETherScan} alt='etherscan link' />
           </div>
         </div>
         {/* footer links */}
